data_drift/requirements.txt
google-cloud-firestore
google-cloud-monitoring